+++ Emulator/FileSystems/FSEmptyBlock.cpp
@@ -10,7 +10,7 @@
 #include "FSDevice.h"
 
 FSItemType
-FSEmptyBlock::itemType(u32 byte) const
+FSEmptyBlock::itemType(isize byte) const
 {
     return FSI_UNUSED;
 }
