+++ Emulator/FileSystems/FSFileHeaderBlock.cpp
@@ -28,14 +28,14 @@ FSFileHeaderBlock(p, nr)
 }
 
 FSItemType
-FSFileHeaderBlock::itemType(u32 byte) const
+FSFileHeaderBlock::itemType(isize byte) const
 {
     // Intercept some special locations
     if (byte == 328) return FSI_BCPL_STRING_LENGTH;
     if (byte == 432) return FSI_BCPL_STRING_LENGTH;
 
     // Translate the byte index to a (signed) long word index
-    i32 word = byte / 4; if (word >= 6) word -= bsize() / 4;
+    isize word = byte / 4; if (word >= 6) word -= bsize() / 4;
 
     switch (word) {
         case 0:   return FSI_TYPE_ID;
@@ -66,7 +66,7 @@ FSFileHeaderBlock::itemType(u32 byte) const
 }
 
 ErrorCode
-FSFileHeaderBlock::check(u32 byte, u8 *expected, bool strict) const
+FSFileHeaderBlock::check(isize byte, u8 *expected, bool strict) const
 {
     /* Note: At locations -4 and -3, many disks reference the bitmap block
      * which is wrong. We ignore to report this common inconsistency if
@@ -74,7 +74,7 @@ FSFileHeaderBlock::check(u32 byte, u8 *expected, bool strict) const
      */
 
     // Translate the byte index to a (signed) long word index
-    i32 word = byte / 4; if (word >= 6) word -= bsize() / 4;
+    isize word = byte / 4; if (word >= 6) word -= bsize() / 4;
     u32 value = get32(word);
     
     switch (word) {
