+++ Emulator/FileSystems/FSRootBlock.cpp
@@ -29,13 +29,13 @@ FSRootBlock::~FSRootBlock()
 }
 
 FSItemType
-FSRootBlock::itemType(u32 byte) const
+FSRootBlock::itemType(isize byte) const
 {
     // Intercept some special locations
     if (byte == 432) return FSI_BCPL_STRING_LENGTH;
 
     // Translate the byte index to a (signed) long word index
-    i32 word = byte / 4; if (word >= 6) word -= bsize() / 4;
+    isize word = byte / 4; if (word >= 6) word -= bsize() / 4;
     
     switch (word) {
         case 0:   return FSI_TYPE_ID;
@@ -69,10 +69,10 @@ FSRootBlock::itemType(u32 byte) const
 }
 
 ErrorCode
-FSRootBlock::check(u32 byte, u8 *expected, bool strict) const
+FSRootBlock::check(isize byte, u8 *expected, bool strict) const
 {
     // Translate the byte index to a (signed) long word index
-    i32 word = byte / 4; if (word >= 6) word -= bsize() / 4;
+    isize word = byte / 4; if (word >= 6) word -= bsize() / 4;
     u32 value = get32(word);
     
     switch (word) {
